add tainttoleration plugin

+++ kube-scheduler/src/main/java/com/lgypro/j8s/framework/ScorePlugin.java
@@ -8,11 +8,19 @@ import com.lgypro.j8s.context.Context;
  * nodes that passed the filtering phase.
  */
 public interface ScorePlugin extends Plugin {
-    long maxNodeScore = 100;
-    long minNodeScore = 0;
-    long maxTotalScore = Long.MAX_VALUE;
+    int maxNodeScore = 100;
+    int minNodeScore = 0;
+    int maxTotalScore = Integer.MAX_VALUE;
 
+    /**
+     * Score is called on each filtered node. It must return success and an integer
+     * indicating the rank of the node. All scoring plugins must return success or
+     * the pod will be rejected.
+     */
     Object[] score(Context context, CycleState state, Pod pod, String nodeName);
 
+    /**
+     * ScoreExtensions returns a ScoreExtensions interface if it implements one, or nil if does not.
+     */
     ScoreExtensions scoreExtensions();
 }
